refactor(test): Move test_empty_expr.py into test_treegen.py
