--- test/test_cstgen/.snapshots/test_empty_expr.txt
@@ -1,15 +1,0 @@
-"2"
-"MyTestCase::test_empty_expr_issue_04:0"
-ProgramNode(StrRegion(0, 9), [
-  DeclNode(StrRegion(0, 7), [
-    DeclScope_Let(StrRegion(0, 3)),
-    DeclType_Variable(StrRegion(3, 3)),
-    DeclItemsList(StrRegion(4, 7), [
-      DeclItemNode(StrRegion(4, 7), [
-        IdentNode(StrRegion(4, 5)),
-        NumberNode(StrRegion(6, 7))
-      ])
-    ])
-  ]),
-  NopNode(StrRegion(8, 9))
-])
